+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ENHANCEMENTS:
 
+- Use LocalUserCreatedHook instead of AddNewAccountHook
 - Add edit-count achievement and $wgAchievementBadgesReplaceEchoThankYouEdit for enabling it.
 - Introduce the types of achievements. 'instant' and 'stats' is available.
 - Add SpecialAchievementsBeforeGetEarned Hook.